+++ playground/DryIoc.SpeedTestApp/Tests/ClosureFieldsAccessSpeed.cs
@@ -9,8 +9,6 @@ using NUnit.Framework;
 
 namespace DryIoc.UnitTests.Net40.Playground
 {
-    //[TestFixture]
-    //[Ignore]
     public class ClosureFieldsAccessSpeed
     {
         public static void TestExpr()